refactor: swappage
