+++ pages/swap-nft/ConfirmedSwap.tsx
@@ -1,7 +1,8 @@
 import React from 'react';
 import { Button } from '../../components/ui/Button';
-import { IStepsProps } from './types';
-import styles from './swap-page.module.scss';
+import { IStepsProps } from '.';
+import styles from '../../styles/pages/swap-page.module.scss';
+
 const ConfirmedSwap = ({ setSteps }: IStepsProps) => (
   <div className={styles.swapSubPage}>
     <p className="font-poppins dark:text-white text-nft-black-1 font-semibold text-xl">
