--- pages/swap-nft/types.d.ts
@@ -1,5 +1,0 @@
-export interface IStepsProps {
-  setSteps: (steps: Steps) => void;
-}
-
-export type Steps = 'WalletAddress' | 'EnterNFTS' | 'ConfirmSwap' | 'Confirmed';
